Move github merge requests operations to github/merge_request
